+++ packages/sdk-v2/src/index.ts
@@ -1,3 +1,4 @@
-export * from './contexts';
-export * from './hooks';
-export * from './sdk';
+export * from "./contexts";
+export * from "./hooks";
+export * from "./sdk";
+export * from "./utils";
