+++ packages/sdk-v2/src/sdk/ipfs/index.ts
@@ -1,2 +1,1 @@
 export * from "./sdk";
-export * from "./utils";
