+++ packages/sdk-v2/src/sdk/ipfs/sdk.ts
@@ -1,12 +1,21 @@
-import { sortBy } from "lodash";
+import { isString, sortBy } from "lodash";
+import { CID } from "multiformats/cid";
 
-import { fallback, toV1, withTemporaryFile } from "./utils";
+import { fallback, withTemporaryFile } from "../../utils";
 
 interface RequestOptions {
   headers?: any;
   body?: any;
   mode?: any;
 }
+
+// eslint-disable-next-line
+export const toV1 = cid => CID.parse(cid).toV1().toString();
+//todo: move to utils
+export const cidRegexp = /^[\w\d]+$/i;
+
+// checks is string a valid CID. it should be at least 40 chars length and contrain only letters & numbers
+export const isValidCID = source => isString(source) && source.length >= 40 && cidRegexp.test(source);
 
 export type IPFSUrls = { ipfsGateways: string; ipfsUploadGateway: string };
