--- packages/sdk-v2/src/sdk/ipfs/utils/toV1.ts
@@ -1,3 +1,0 @@
-import { CID } from "multiformats/cid";
-// eslint-disable-next-line
-export const toV1 = cid => CID.parse(cid).toV1().toString();
