+++ packages/sdk-v2/src/sdk/newsfeed/index.ts
@@ -1,2 +1,2 @@
-export * from "./sdk";
+export * from "./OrbisCachedFeed";
 export * from "./utils";
