--- packages/sdk-v2/src/sdk/newsfeed/sdk.ts
@@ -1,1 +1,0 @@
-export { OrbisCachedFeed } from "./OrbisCachedFeed";
