+++ packages/sdk-v2/src/sdk/newsfeed/utils/index.ts
@@ -1,2 +1,1 @@
 export * from "./createDb";
-export * from "./feedUtils";
