+++ packages/sdk-v2/src/utils/base64.ts
