+++ packages/sdk-v2/src/utils/fs.ts
