+++ packages/sdk-v2/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from "./async";
 export * from "./base64";
 export * from "./fs";
-export * from "./toV1";
